jdbc: fix hardcoded driver version
Use build-helper-maven-plugin to parse ${project.version} and
templating-maven-plugin to substitute placeholders in Version.java.

It is possible that IntelliJ IDEA will not detect generated sources
automatically, shows 'cannot resolve symbol' errors re Version class and
even fails to build the project. In this case execute `mvn clean
compile` in the project root from a terminal, switch to IDEA, click
right button on the project in the left-hand side panel and choose Maven
-> Reimport.

There is also the relevant [issue][1] in the JetBrains tracker.

[1]: https://youtrack.jetbrains.com/issue/IDEA-161010