test: add MsgPackLite test cases
We are using a borrowed implementation of the MsgPack library w/o
available tests. In addition, there are a few modifications that were
done on our side. It requires the tests that will cover at least typical
use cases related to (de-)serialization between designated java types
and message pack types.

This commit introduces a set of test cases that check correctness of
packing/unpacking according to the MsgPack specification.

Closes: #127