iproto: add PREPARE operation
Tarantool introduced a new PREPARE protocol operation to be able to
create a prepared statement on the server side. This can be used by
JDBC SQLPreparedStatement implementation that will prepare its
query in advance to get meta data as well as reuse it multiple times.
These JDBC features will be introduced in future commits.

This commit add all the required protocol constants and allow
TarantoolClient to abstract from concrete operation codes related to
SQL. Because the native client does not provide API to perform PREPARE
it does not handle it in completeSql() as a completely unexpected
operation here.

Follows on: #198