jdbc: support ParameterMetaData class
In addition to result set metadata it's possible to examine parameters
of PreparedStatement using getParameterMetaData() method. Because
Tarantool returns extra info related to query parameters as a result of
PREPARE operation, we can fill ParameterMetaData by available info.

However, the server sends always 'ANY' as a target parameter type for
parameters and the driver treats all of them as UNKNOWN type. Once the
server starts to send proper types (such as integer, string and so on)
the driver should parse it automatically (required to be tested in
future).

Follows on: #173