metadata: Tarantool instance version info
The driver can connect to different Tarantool instances within a range
of server versions. Sometimes, it's required to set a driver behaviour
depending on those versions (i.e. in scope of #213 the driver needs to
generate the functions for JDBC C4 LTRIM/RTRIM that are recognizable by
a particular Tarantool version).

This commit also includes implementation of two public DatabaseMetaData
methods getDatabaseMajorVersion and getDatabaseMinorVersion.

Affects: #213
Closes: #106